Type corrections
status should be signed, as the return codes are signed.
Pointers in the ELF code are not deltas, so they should be unsigned.
The SMC interface takes unsigned values, so we add an explicit cast.

Bug: 129300035

+++ lib/sm/smcall.c
@@ -134,7 +134,7 @@ static long smc_cpu_resume(smc32_args_t* args) {
 
 #if WITH_LIB_VERSION
 static long smc_get_version_str(smc32_args_t* args) {
-    int32_t index = args->params[0];
+    int32_t index = (int32_t)args->params[0];
     size_t version_len = strlen(lk_version);
 
     if (index == -1)
